+++ package.json
@@ -11,6 +11,7 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-icons": "^5.5.0",
+    "react-router-dom": "^6.30.0",
     "react-scripts": "^5.0.1",
     "react-tooltip": "^5.28.1",
     "web-vitals": "^2.1.4"
