+++ src/login.css
@@ -1,0 +1,91 @@
+/* General Styles */
+body {
+  margin: 0;
+  padding: 0;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  background-color: #181818;
+  font-family: Arial, sans-serif;
+}
+
+/* Login Container */
+.login-container {
+  width: 350px;
+  background-color: #3a3a3a;
+  border-radius: 10px;
+  overflow: hidden;
+  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.3);
+}
+
+/* Login Box */
+.login-box {
+  padding: 30px;
+  background-color: #3a3a3a;
+  display: flex;
+  flex-direction: column;
+}
+
+/* Login Header */
+.login-header {
+  background: linear-gradient(to right, #2f8f63, #4db8ff);
+  color: white;
+  padding: 15px;
+  font-size: 16px;
+  font-weight: bold;
+  text-align: left;
+}
+
+/* Input Fields */
+.login-box input {
+  width: 100%;
+  padding: 10px;
+  margin: 10px 0;
+  border: none;
+  border-radius: 5px;
+  background-color: #252525;
+  color: white;
+  font-size: 14px;
+}
+
+.login-box input::placeholder {
+  color: #aaa;
+}
+
+/* Login Button */
+.login-box .login-btn {
+  width: 100%;
+  padding: 10px;
+  background: linear-gradient(to right, #2f8f63, #4db8ff);
+  color: white;
+  border: none;
+  border-radius: 5px;
+  font-size: 16px;
+  cursor: pointer;
+  transition: 0.3s;
+  margin-top: 10px;
+}
+
+.login-box .login-btn:hover {
+  background: linear-gradient(to right, #4db8ff, #2f8f63);
+}
+
+/* Sign-Up Button */
+.login-box .signup-btn {
+  width: 100%;
+  padding: 10px;
+  background: transparent;
+  border: 2px solid white;
+  color: white;
+  border-radius: 5px;
+  font-size: 16px;
+  cursor: pointer;
+  transition: 0.3s;
+  margin-top: 10px;
+}
+
+.login-box .signup-btn:hover {
+  background: white;
+  color: #181818;
+}
